⚡ Cut CI wall clock and fix the red develop build
The style gate scanned files one process at a time, the Nx cache never survived
a run, Storybook sat serially inside the test job, and every smoke image rebuilt
its dependencies from scratch.

- agent-style-check: one node/grep/awk pass over the whole file list instead of
  ~9 processes per file (1101 files: ~7min -> 36s, byte-identical findings).
- Exclude vendored .upstream/ trees from the style rules, and split the one real
  double-ternary in runtime-bootstrap.router.test.ts.
- Move the MCP wire shapes into their paired mcp-gateway.types.ts.
- Nx cache: restore before the first Nx task and save with if: always(), so a run
  that fails late no longer leaves the next one cold.
- Storybook: build-storybook/test-storybook/test-storybook-visual are cacheable,
  and their inputs come from one storybookCatalogue named input rather than a
  hand-kept package list that had already drifted by one package.
- Storybook moves to its own job, parallel to test; both publish jobs still gate
  on it.
- Drop the redundant setup-node npm cache (node_modules is cached directly), which
  frees ~2GB against the repo's 10GB cache ceiling.
- Smoke images build a few at a time instead of strictly one after another.
- Dockerfiles copy the install manifests before npm ci and the sources after, so
  editing a library no longer invalidates the dependency layer.
