+++ apps/artifact-scanner/deploy/Dockerfile
@@ -2,9 +2,9 @@
 FROM node:22-bookworm-slim@sha256:d649c27dae7ba0137b3cef5dd75baa422c08dc3d9e3fc0c23dfb172dc3cc6436 AS build
 WORKDIR /app
 COPY package.json package-lock.json tsconfig.json nx.json ./
-COPY libs libs
 COPY apps/artifact-scanner/package.json apps/artifact-scanner/project.json apps/artifact-scanner/tsconfig.json apps/artifact-scanner/
 RUN npm ci
+COPY libs libs
 COPY apps/artifact-scanner/src apps/artifact-scanner/src
 RUN npx nx run artifact-scanner:build
