+++ apps/artifact-service/deploy/Dockerfile
@@ -3,9 +3,9 @@ FROM node:22-bookworm-slim AS build
 
 WORKDIR /app
 COPY package.json package-lock.json tsconfig.json nx.json ./
-COPY libs libs
 COPY apps/artifact-service/package.json apps/artifact-service/project.json apps/artifact-service/tsconfig.json apps/artifact-service/
 RUN npm ci
+COPY libs libs
 
 COPY apps/artifact-service/src apps/artifact-service/src
 RUN npx nx run artifact-service:build
