+++ libs/frontend/elements/ui/project.json
@@ -30,16 +30,7 @@
 			"inputs": [
 				"default",
 				"^production",
-				"{workspaceRoot}/libs/frontend/core/src/lib/theme/opencrane-theme.scss",
-				"{workspaceRoot}/libs/frontend/elements/a2ui/src/**/*",
-				"{workspaceRoot}/libs/frontend/features/onboarding/src/**/*",
-				"{workspaceRoot}/libs/frontend/features/context/src/**/__tests__/*.stories.ts",
-				"{workspaceRoot}/libs/frontend/features/context/src/lib/components/canvas-doc/**/*",
-				"{workspaceRoot}/libs/frontend/features/conversation-assets/src/**/*",
-				"{workspaceRoot}/libs/frontend/features/conversation-activity/src/**/*",
-				"{workspaceRoot}/libs/frontend/features/conversation-elicitation/src/**/*"
-				,"{workspaceRoot}/libs/frontend/elements/conversation/src/**/*"
-				,"{workspaceRoot}/libs/frontend/features/agent-threads/src/**/*"
+				"storybookCatalogue"
 			],
 			"options": {
 				"browserTarget": "frontend-elements-ui:build-storybook",
@@ -62,19 +53,11 @@
 		},
 		"build-storybook": {
 			"executor": "@storybook/angular:build-storybook",
+			"cache": true,
 			"inputs": [
 				"default",
 				"^production",
-				"{workspaceRoot}/libs/frontend/core/src/lib/theme/opencrane-theme.scss",
-				"{workspaceRoot}/libs/frontend/elements/a2ui/src/**/*",
-				"{workspaceRoot}/libs/frontend/features/onboarding/src/**/*",
-				"{workspaceRoot}/libs/frontend/features/context/src/**/__tests__/*.stories.ts",
-				"{workspaceRoot}/libs/frontend/features/context/src/lib/components/canvas-doc/**/*",
-				"{workspaceRoot}/libs/frontend/features/conversation-assets/src/**/*",
-				"{workspaceRoot}/libs/frontend/features/conversation-activity/src/**/*",
-				"{workspaceRoot}/libs/frontend/features/conversation-elicitation/src/**/*"
-				,"{workspaceRoot}/libs/frontend/elements/conversation/src/**/*"
-				,"{workspaceRoot}/libs/frontend/features/agent-threads/src/**/*"
+				"storybookCatalogue"
 			],
 			"outputs": [
 				"{options.outputDir}"
@@ -114,7 +97,12 @@
 		},
 		"test-storybook": {
 			"executor": "nx:run-commands",
-			"cache": false,
+			"cache": true,
+			"inputs": [
+				"default",
+				"^production",
+				"storybookCatalogue"
+			],
 			"dependsOn": [
 				"static-storybook"
 			],
@@ -124,15 +112,14 @@
 		},
 		"test-storybook-visual": {
 			"executor": "nx:run-commands",
-			"cache": false,
+			"cache": true,
 			"dependsOn": [
 				"static-storybook"
 			],
 			"inputs": [
 				"default",
 				"^production",
-				"{workspaceRoot}/playwright.storybook.config.ts",
-				"{workspaceRoot}/tests/storybook/**/*"
+				"storybookCatalogue"
 			],
 			"outputs": [
 				"{workspaceRoot}/.nx/test-results/storybook"
